+++ bauh/gems/flatpak/flatpak.py
@@ -8,6 +8,7 @@ from bauh.api.exception import NoInternetException
 from bauh.commons.system import new_subprocess, run_cmd, new_root_subprocess
 
 BASE_CMD = 'flatpak'
+RE_SEVERAL_SPACES = re.compile(r'\s+')
 
 
 def get_app_info_fields(app_id: str, branch: str, fields: List[str] = [], check_runtime: bool = False):